feat: Add compatibility with core Lightning wire types
This change allows us to be compatible with core Lightning wire
types. In turn, core Lightning should aim to be compatible with
the specification type.

Currently, core Lightning uses `u8` for the `bytes` alias and `amount_msat`
for `u64`. We've addressed this issue in the parser for now, but we
should be able to implement the transformation in the AST later.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>